Added a check so we can't delete our user that is currently logged in.

+++ NackademinWebShop/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
@@ -63,7 +64,14 @@ namespace NackademinWebShop.Controllers
         //TODO FIX ME:
         public async Task<IActionResult> Delete(string id)
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            if (userId == id)
+            {
+                ModelState.AddModelError("", "You can't delete your own user.");
+                return RedirectToAction("GetAll");
+            }
             var user = _userManager.Users.FirstOrDefault(p => p.Id == id);
+
             var result = await _userManager.DeleteAsync(user);
 
             return RedirectToAction("GetAll");
